requirements_airflow.txt
apache-airflow-providers-postgres
pandas
yfinance
sqlalchemy
psycopg2-binary
pymysql